# Add inline attributes to Vec's MutBuf methods